src/main/java/org/example/problem35/Solution.java
package org.example.problem35;

public class Solution {
    public int searchInsert(int[] nums, int target) {
        int low = 0;
        int high = nums.length;
        if(target > nums[high-1]){
            return high;
        }
        if(target < nums[low]){
            return low;
        }
        while (low <= high) {
            int mid = low + (high - low) / 2;
            if (nums[mid] < target) {
                low = mid + 1;
            } else if (nums[mid] > target) {
                high = mid - 1;
            } else if (nums[mid] == target) {
                return mid;
            }
        }
        return low;
    }

}
